Move crypto functions from utils to crypto
