+++ crypto/hash_test.go
@@ -1,4 +1,4 @@
-package utils
+package crypto
 
 import (
 	"github.com/stretchr/testify/assert"
